guitara/requirements.txt
# Django and REST framework
django==5.1.4
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-rest-knox==4.2.0
django-filter==23.5

# Fixing dependency issue with cryptography
cryptography==40.0.2

# Real-time functionality
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
redis==5.0.1

# Database
psycopg2-binary==2.9.9

# Password hashing
bcrypt==4.1.2

# Testing and development
pytest==8.0.0
pytest-django==4.8.0
coverage==7.4.1

# Utility
python-dotenv==1.0.0
